+++ app/src/main/java/com/developers/sprintsync/home/statistics/domain/model/TracksStatsCompact.kt
@@ -2,9 +2,9 @@ package com.developers.sprintsync.home.statistics.domain.model
 
 data class TracksStatsCompact(
     val runs: Int,
-    val totalDistance: Int,
+    val totalDistance: Float,
     val totalCalories: Int,
-    val longestDistance: Int,
+    val longestDistance: Float,
     val bestPace: Float,
     val maxDuration: Long,
 ) {
@@ -12,9 +12,9 @@ data class TracksStatsCompact(
         val EMPTY =
             TracksStatsCompact(
                 runs = 0,
-                totalDistance = 0,
+                totalDistance = 0f,
                 totalCalories = 0,
-                longestDistance = 0,
+                longestDistance = 0f,
                 bestPace = 0f,
                 maxDuration = 0L,
             )
