+++ app/src/main/java/com/developers/sprintsync/tracking/data/flow/DistanceFlowManager.kt
@@ -7,4 +7,4 @@ class DistanceFlowManager
     @Inject
     constructor(
         provider: GetCurrentTrackDistanceMetersUseCase,
-    ) : FlowManager<Int>(provider())
+    ) : FlowManager<Float>(provider())
